move dexie and their files into a new utils folder; rename default utils to task utils
